Passed all gradescope test cases

+++ cache.cpp
@@ -643,29 +643,29 @@ int cache::findEvictL2Index(int index)
 }
 
 // Calculate L1 miss rate
-float cache::calcL1MissRate()
+double cache::calcL1MissRate()
 {
-	float res;
+	double res;
 	// cout << myStat.missL1 << endl;
 	// cout << myStat.accL1 << endl;
-	res = ((static_cast<float>(myStat.missL1)) / (myStat.missL1 + myStat.accL1));
+	res = ((static_cast<double>(myStat.missL1)) / (myStat.missL1 + myStat.accL1));
 	// cout << res << endl;
 	return res;
 }
 
 // Calculate L2 miss rate
-float cache::calcL2MissRate()
+double cache::calcL2MissRate()
 {
-	float res;
-	res = ((static_cast<float>(myStat.missL2)) / (myStat.missL2 + myStat.accL2));
+	double res;
+	res = ((static_cast<double>(myStat.missL2)) / (myStat.missL2 + myStat.accL2));
 	return res;
 }
 
 // Calculate VC miss rate
-float cache::calcVCMissRate()
+double cache::calcVCMissRate()
 {
-	float res;
-	res = ((static_cast<float>(myStat.missVic)) / (myStat.missVic + myStat.accVic));
+	double res;
+	res = ((static_cast<double>(myStat.missVic)) / (myStat.missVic + myStat.accVic));
 	return res;
 }
