+++ cache.h
@@ -76,9 +76,9 @@ public:
 	bool L2Full(int index);
 	int findEvictL2Index(int index);
 
-	float calcL1MissRate();
-	float calcL2MissRate();
-	float calcVCMissRate();
+	double calcL1MissRate();
+	double calcL2MissRate();
+	double calcVCMissRate();
 
 	// bool updateMainMem(int *data, int *myMem); // updates data in Main Memory
 };
